requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy
pymysql
cryptography
google-genai
passlib
python-jose[cryptography]
email-validator
